Create php.ini
Std. production php.ini, with Xdebug addition in bottom. On a standard Windows 10 x64 installation, the xdebug.remote_host is set to 10.0.75.1. If this does not work, run cmd.exe and ipconfig. You will need to use the IPv4 Address in (DockerNAT), which i guess would default to 10.0.75.1 on most newer systems (windows based)

xdebug.remote_connect_back=1 does NOT work in docker when using linked containers. You WANT the linked containers, and therefore you have to rely on remote_host (no opening/exposing of port 9000 is required, as the debugger (phpstorm?) will act as server)